requirements.txt
groq
tweepy
python-dotenv